Move feeds tests into seperate files/dir
